scripts/requirements.txt
termcolor~=2.4.0
opencv-python
Pillow~=9.0.1
openai
pillow-avif-plugin